% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{card}
\alias{card}
\title{Create card}
\usage{
card(categoryname, title, imgUrl, dependency = NULL)
}
\arguments{
\item{dependency}{default=NULL}
}
\description{
Create card
}
\examples{
\dontrun{
card(
  categoryname="臺北大學城",
  title="【USR】社科院USR專書出版-「大學責任與高齡學習」",
  imgUrl ="https://miro.medium.com/max/1400/1*2v4yO6cR2XOSlI65Ge6GrA.jpeg"
) |> browseTag2()}
}
