% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poster.R
\name{poster}
\alias{poster}
\title{Poster UI}
\usage{
poster(
  title = "USR公民培訓課程：UI設計神器Figma",
  date = "4.26, 4.29",
  imgSrc = "lib/usrApp-1.0.0/img/figmaposter-sm.jpg",
  context = "test",
  dependency = NULL
)
}
\arguments{
\item{dependency}{}
}
\description{
Poster UI
}
\examples{
none.
}
