% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{scale_poster}
\alias{scale_poster}
\title{Scale poster image proportionally to width 360px}
\usage{
scale_poster(filename)
}
\arguments{
\item{filename}{}
}
\value{

}
\description{
Scale poster image proportionally to width 360px
}
\examples{
none.
}
